requirements.txt
# SnapEats - Root Dependencies

# AI/ML
ultralytics==8.2.0
torch==2.4.0
torchvision==0.19.0
opencv-python==4.10.0.84
numpy==1.26.4
Pillow==10.3.0

# Web Framework & Backend
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
python-dotenv==1.0.0

# Security
cryptography==42.0.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1

# Database
sqlalchemy==2.0.27

# Data Validation
pydantic==2.6.1

# Configuration
pyyaml==6.0.1

# Testing (Optional)
pytest==7.4.4
pytest-asyncio==0.23.3

# Development Tools
black==24.1.1
flake8==7.0.0
